api(dailygit): modified deleteFolder mutation
